test(mcp): increase code coverage
